Update binary_sensor.ffmpeg.markdown (#886)
Reworded a few places for grammar/readability.

Also, I have never proposed a change to the documentation before, so I hope I am doing this in the proper way.  Trying to be helpful, not step on anyone's feet!  Please feel free to admonish me, if I am not following proper etiquette. :)