Update installation-raspberry-pi-image.markdown (#1459)
* Update installation-raspberry-pi-image.markdown

Add info to customize. I had to search a few minutes to figure out how to change the keyboard layout. 

Add Windows 10 SSH in bash info. Maybe there should be a seperate page in the documentation, where all 'How to connect/SSH' topics link to. That way it won't have to be repeated throughout the docs.

* Update installation-raspberry-pi-image.markdown

* Update installation-raspberry-pi-image.markdown

Reworked mention of Windows 10 bash feature mention.

* Update installation-raspberry-pi-image.markdown

Contained a typo, but also changed the wording