Update binary_sensor.xiaomi.markdown (#3094)
I would recommend to turn the automation off/on for turning off the Lights because, if you do not do this, and you for example turn manually the light on and walk into your motion sensor it will turn off the light(s). I have had this a few times.