Mochad light fixes (#4214)
* Add brighness levels option to mochad

In home-assistant/home-assistant#11146 a new configuration option,
brighness_levels, is added to the mochad light component. This is used
to tell home-assistant how many brightness levels the X10 light device
supports because there is no way to actually discover this. This commit
adds documentation for the new option.

* Fix some leftover references to the switch component

The documentation for the mochad light component was mostly copy and
pasted from the mochad switch device. Since the configuration is more or
less the same between devices. However in doing this several references
to mochad switches was left in by mistake. This commit fixes these and
replaces them with the correct name of the component.