sync usb_path to other docs (#4439)
* sync usb_path to other docs

The hassio zwave documents reference usb_path: /dev/ttyACM0, so this document should be in sync. After some discussion it appears that this could be variable, in which case there should be a good way to figure out the CORRECT answer rather than using a guess as example code without stating it as such. Possibly linking back to the linux "how to find" section?

* Removed manual set language per suggestion