Add some emphasis about Breaking Changes
We should provide proper guidance to the user about the importance of release notes and paying attention to breaking changes.  A common error users face while updating home assistant is not paying attention to the breaking changes.  If the user has not done an update in sometime they may also forget about previous breaking changes that still stand.