Adds comment about deprecated homeassistant_start event
The note in the documentation for triggers says that the
homeassistant_start event has been deprecated. This PR just adds that
note to the events page.

It is unclear to me if homeassistant_stop has also been replaced by
'stop' on the 'homeassistant' platform. If so, the note should probably
be expanded on both pages to indicate that.