Update description about API Key
We should recommend users to setup this optional configuration key, as some were unaware that this is a required step when they add devices in the future so we should inform them of this.  Some users are not aware that they can also issue a voice command to sync the home assistant device list.