Added port 3307 to Maira DB (#4743)
Maria DB 10 have changed port from 3306 that Maria DB are using and is standard in this component. Maria DB is using port 3307 instead and that need to be added in the configuration to work. I have added a note in the description to add the port to the server_ip.