Add documentation for the new watson_iot component (#5100)
* Add documentation for the new watson_iot component

This commit adds documention for the new watson_iot component which was
added in home-assistant/home-assistant#13664
It links to the official documentation for setting up the cloud service
and generating the required auth information.

* :pencil2: Missing comma

* Use configuration tags

This commit updates the formatting of the configuration section of the
documentation. Previously it manually formatted a list instead of
leveraging the templating format. This corrects that oversight and uses
the standard format.

* :art: Adds logo

* :arrow_up: ha_release -> 0.68

* :pencil2: Processed RFC's