Update mailgun.markdown (#6143)
Fix inconsistency where "domain" is required but "sandbox" can also be used to specify the domain.  It appears that the package https://github.com/pschmitt/pymailgunner in  def guess_domain allows multiple domains and "sandbox" can be used to choose a sandbox domain over a normal domain.  As far as I can tell, this feature isn't used by the Mailgun component.  There are also updates to the Mailgun Notify doc.