Add second computer and setup note (#6478)
* Add second computer and setup note

Added a second computer to the config example as I was stumped for a bit on how to get a second one entered.  Its obvious now but has stumped others too based on forum searches.  

The note is another hiccup some may run into.  I listed the problem and promoted the user to search for the windows settings changes.  I assume HomeAssistant doesn't want to specify firewall and registry settings to be made.

* Fix a typo and update the style

* :pencil2: Spelling and grammar