The current text assumes your user is always "homeassistant" (#7642)
* The current text assumes your user is always "homeassistant"

If you follow the advanced guide to setup a virtual environment a user named "homeassistant" or whatever you want will be created. See:
https://www.home-assistant.io/docs/installation/raspberry-pi/
If you follow this guide the "homeassistant" user won't be created:
https://www.home-assistant.io/docs/installation/virtualenv
This page actually refers to the Autostart page on which is assumed that you created a user named "homeassistant". It's better to refer to the user which runs HA in general.

* All occurences of [your user] replaced by YOUR_USER

* Update systemd.markdown