Fedora improvements (#7774)
* Remove group gpio from Installation instructions for Fedora

That group doesn't exist on Fedora by default. I believe it makes more
sense to add homeassistant to that group later when that group needs to
be created.

* sudo to homeassistant before virtualenv is activated on Fedora

Very similar approach to Debian-based installation needs to be done
here. In the previous block of commands user homeassistant is created
and directories are chowned to him.

If one would try to run followup pip3 install or even hass itself as
current user he wouldn't be able to write to those directories or that
homeassistant user would be just useless.