Added configuration when using separate sensors per tariff (#8439)
* Added configuration when using separate sensors per tariff

Additional configuration example and explanation added for DSMR users, or other setups that have separate source sensors to track for different power tariffs.

* note created about persistent sensor values

Paragraph about persistent values for this component are now in a note css

* Remove info overdose 

No need to explain that 6 new entities are created with the example, as it is clear from the yaml that 6 entities are configured. Removed.