Clarify which password to use with velux component (#8706)
* Clarify which password to use with velux component

Previous examples everywhere use the "velux123" string. That is the web login password the KLF 200 defaults to. The password actually required is the other one, which is both significantly more complex and set elsewhere.
Changing the example will hopefully help others to not waste as much time on this as I did ;-)

* Minor change as stated in the guidelines