Additional clarification (#8900)
* Additional clarification

I was a bit confused about what the base url was supposed to be here and that setting it here was overriding the HTTP setting in configuration.yaml.  Often I see in the documentation that there are examples without explanation of why that was used as the example.  So for this one, I assumed that it was `asking` for the base_url that I had set, rather than allowing me to define a custom base url.

* :pencil2: Tweak