Minor code formatting fixes in Owntracks document (#9397)
This fixes a bug with the code formatting in a few places of the owntracks document. Several times, example values are in the document using the format "<text>". I believe that markdown is treating these as HTML tags, when really they are not. As a result, some of the text inside of these tags are not displayed.

This problem can be seen here: https://www.home-assistant.io/components/owntracks/#configuring-the-app---android

Several of these issues have been fixed by using markdown code formatting around the example values.