Add tunnel timeout to support websocket connections (#11708)
HomeAssistant uses a 55 second interval[1] to send heartbeats to
the Lovelace UI over WebSockets. If the tunnel timeout[2] is not
set, the WebSocket connection from the browser to the server is
reset every 50 seconds. When the connection resets, it reconnects
automatically, and causes the page to reload. Any unsaved data in
forms is lost.

This commit adds a sane default of 60 seconds to the HAProxy
configuration for tunneled connections, so that the frontend
WebSocket connection doesn't time out every 50 seconds.

[1]: https://github.com/home-assistant/home-assistant/blob/9c551ae85d68a8c82c706d6e2cf044cfd5533054/homeassistant/components/websocket_api/http.py#L111
[2]: https://cbonte.github.io/haproxy-dconv/1.8/configuration.html#4-timeout%20tunnel

Related: home-assistant/home-assistant-js-websocket#108