move note about check config (#11939)
Moves the note about the "CHECK CONFIG" button only being visible in advanced mode to the previous paragraph where it tells you to click the button. I was confused for a second until I read on to the next paragraph. It seems to make more sense to put the note about the button in the paragraph where it tells you to click the button.

I also made a couple other minor tweaks for consistency and accuracy.