Update google_assistant.markdown (#14529)
If the user is not in the correct project when they create the service account this will not work properly. Update the steps associated with getting project ids to match. This can now only be done by selecting the Actions on Google project_id from within Google Cloud API Console.