Fix Link (#14891)
The link was broken for this page so just a simple fix.  There were a few API documentation on the Pilight website, but after reading the Home Assistant documentation, I was under the impression that the Socket API link was the proper fit.  However, I could be mistaken, so any feedback would be appreciated.