Update velux.markdown (#16051)
* Update velux.markdown

Added description of how to connect the new Velux Active gateway to Home Assistant. It doesn't have it's own API, but is working perfectly through the HomeKit Controller integration. I know many people have asked about this and I thought it would be good to include the information in the Velux integration documentation.
Maybe this information should be marked with the "note" tag?

* Update source/_integrations/velux.markdown

Absolutely!

Co-authored-by: Franck Nijhof <frenck@frenck.nl>

Co-authored-by: Franck Nijhof <frenck@frenck.nl>