Fix the display of default values (#18094)
The `{% configuration %}` section uses yaml and some jekyll template to generate the html output of the configuration description. However, yaml parses the raw values `ON` and `OFF` as booleans, which are then displayed as `true` and `false` in the final rendered html. This change adds quotes around those values so that they are parsed as strings and displayed as is in the html.